Modularize db/issues/tests.rs into submodules for maintainability
Split 641-line tests.rs into domain-focused submodules:
- crud.rs: save/load operations
- labels.rs: label parsing
- filters.rs: basic filtering
- sorting.rs: sort order tests
- duration.rs: duration parsing
- issue_filter.rs: IssueFilter struct tests
